+++ tap_facebook/streams/campaign.py
@@ -127,7 +127,6 @@ class CampaignStream(IncrementalFacebookStream):
         Property("ad_strategy_id", IntegerType),
         Property("lifetime_budget", StringType),
         Property("last_budget_toggling_time", StringType),
-        Property("daily_budget", IntegerType),
         Property("special_ad_category_country", ArrayType),
     ).to_dict()